backend/requirements.txt
# Web server
flask>=2.0
flask-cors>=3.0

# Image + numeric processing
Pillow>=9.0
numpy>=1.24

# HTTP client for smoke tests
requests>=2.28

# Optional: production WSGI server
gunicorn>=20.0
psycopg2-binary==2.9.9